refactor: move hook to lib/auth
