+++ apps/web/src/app/signup/CodeAuthForm.tsx
@@ -5,59 +5,20 @@ import { Form, FormField, FormItem, FormLabel, FormControl, FormMessage } from "
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as z from "zod";
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { Button } from "@paperplane/ui/button";
 import { Loader2, RocketIcon } from "lucide-react";
 import { Input } from "@paperplane/ui/input";
 import { useRouter } from "next/navigation";
 import { CodesDialog } from "./CodesDialog";
-import { type PaperPlaneApiOutputs, api } from "#trpc/server";
+import { api } from "#trpc/server";
 import { getTRPCError } from "@paperplane/utils";
+import { UseTwoFactorKey } from "#lib/auth";
 
 export interface AuthFormProps {
 	domains: string[];
 	invite: boolean;
 }
-
-type MFAKeyData = NonNullable<PaperPlaneApiOutputs["v1"]["auth"]["signup"]["mfa"]>;
-
-const UseTwoFactorKey = () => {
-	const [keyData, setKeyData] = useState<MFAKeyData>({ key: "", secret: "", uri: "" });
-
-	/**
-	 * Creates a valid 2fa QR-code
-	 * @param domain The domain which is currently selected
-	 * @returns qr-code image url
-	 */
-	const getImage = (domain: string) => {
-		const BASE_URL = "https://chart.googleapis.com/chart?chs=166x166&chld=L|0&cht=qr&chl=";
-		const imageData = encodeURIComponent(keyData.uri.replace("DOMAIN_PLACEHOLDER", domain));
-
-		return `${BASE_URL}${imageData}`;
-	};
-
-	/** Updates the MFA Key data */
-	const updateMfa = async () => {
-		const data = await api()
-			.v1.auth.signup.mfa.query()
-			.catch(() => null);
-		if (data) setKeyData(data);
-
-		console.debug(`[MFA]: Updated MFA data, data=${data ? "object" : "null"}`);
-	};
-
-	useEffect(() => {
-		void updateMfa();
-
-		const interval = setInterval(updateMfa, 9e5);
-		return () => clearInterval(interval); // Re-generates the 2fa data every time the data has expired
-	}, []);
-
-	return {
-		...keyData,
-		getImage
-	};
-};
 
 export const CodeAuthForm: React.FC<AuthFormProps> = ({ domains, invite }) => {
 	const router = useRouter();
